Rename cert-util.js to ob-util.js
