+++ app/ob-directory.js
@@ -1,5 +1,5 @@
 const request = require('superagent');
-const { setupMutualTLS } = require('./certs-util');
+const { setupMutualTLS } = require('./ob-util');
 const nJwt = require('njwt');
 const qs = require('qs');
 const { session } = require('./session');
